feat: Migrate from Markdown to JSON parser for review results
- Removed the old ReviewResultParser and related imports.
- Simplified BugmentAction to directly use JsonReviewResultParser.
- Eliminated the environment variable for parser selection.
- Enhanced debugging output for better traceability.
- Added IssueUtils and LocationUtils for issue and location handling.
- Updated documentation to reflect the migration and new architecture.
- Ensured all tests passed with the new JSON parsing implementation.
