+++ docs/DEBUG_JSON_PARSER.md
@@ -9,6 +9,7 @@
 当 JSON 解析器运行时，会输出以下调试信息：
 
 ### 1. 原始 LLM 输出
+
 ```
 📝 Raw LLM output (first 500 chars):
 [LLM 输出的前 500 个字符]
@@ -20,6 +21,7 @@
 ```
 
 ### 2. 清理后的内容
+
 ```
 🧹 Cleaned result (first 500 chars):
 [清理后内容的前 500 个字符]
@@ -31,6 +33,7 @@
 ```
 
 ### 3. 错误详情（如果解析失败）
+
 ```
 ❌ Failed to parse JSON review result: [错误信息]
 
@@ -50,28 +53,32 @@
 ## 常见问题排查
 
 ### 问题 1：LLM 输出包含额外内容
+
 **症状**：错误信息显示 "Unexpected non-whitespace character after JSON"
 **解决**：检查 LLM 输出的后 500 字符，看是否在 JSON 后面有额外的文本
 
 ### 问题 2：LLM 输出仍然是 Markdown 格式
+
 **症状**：原始输出显示 Markdown 格式而不是 JSON
 **解决**：检查 prompt.md 模板是否正确更新，或者 LLM 是否需要更多时间适应新格式
 
 ### 问题 3：JSON 格式错误
+
 **症状**：语法错误，如缺少引号、逗号等
 **解决**：查看错误位置的上下文，检查 JSON 格式是否正确
 
-## 临时回退方案
+## 问题解决方案
 
-如果 JSON 解析器持续出现问题，可以临时回退到 Markdown 解析器：
+如果 JSON 解析器出现问题，可以：
 
-```bash
-export BUGMENT_USE_JSON_PARSER=false
-```
+1. 检查 LLM 输出格式是否正确
+2. 查看调试信息定位具体问题
+3. 根据错误信息调整 prompt 模板
 
 ## 调试信息示例
 
 ### 成功解析的示例
+
 ```
 🔍 Starting to parse JSON review result...
 📝 Raw LLM output (first 500 chars):
@@ -95,6 +102,7 @@ export BUGMENT_USE_JSON_PARSER=false
 ```
 
 ### 解析失败的示例
+
 ```
 🔍 Starting to parse JSON review result...
 📝 Raw LLM output (first 500 chars):
@@ -105,7 +113,7 @@ export BUGMENT_USE_JSON_PARSER=false
     ]
   },
   "issues": []
-} 
+}
 
 这是一些额外的文本内容...
 
@@ -113,7 +121,7 @@ export BUGMENT_USE_JSON_PARSER=false
 🔍 JSON Syntax Error Details:
    Error message: Unexpected non-whitespace character after JSON at position 123
    Context around position 123:
-   "} 
+   "}
 
 这是一些额外的文本内容..."
      ^
