+++ src/parsers/index.ts
@@ -3,7 +3,4 @@
  */
 
 export * from "./DiffParser";
-export * from "./ReviewResultParser";
 export * from "./JsonReviewResultParser";
-export * from "./IssueParser";
-export * from "./LocationParser";
